完善视频抓取脚本

+++ sys-grab/grab-stock-python/dongle/video/m3u8.py
@@ -1,18 +1,41 @@
 import requests
 import queue,datetime,threading,os
+import uuid
 
-url = ""
-DIR = r""
-source_file = DIR + "list.txt"
-target_file = DIR + "video.mp4"
-if os.path.exists(DIR) is False:
-    print("Please create a download directory first!")
-    exit(-1)
-isAes = False
+import sys
+
+url = sys.argv[1] or ""
+file_name = sys.argv[2] or "video"
+file_num = sys.argv[3] or "01"
+
+if url == "":
+    raise Exception("Please input url!")
+
+print("receive param: %s %s %s" % (url,file_name,file_num))
+
+DIR = "/data/temp/video/" # 根目录
+TARGET_DIR = DIR + file_name + "/" # 目标目录，mp4存放目录
+if os.path.exists(TARGET_DIR) is False:
+    os.makedirs(TARGET_DIR)
+target_file = "%s%s-%s.mp4" % (TARGET_DIR,file_name,file_num)
+
+import hashlib
+identy = hashlib.md5(url.encode('utf-8')).hexdigest()
+TMP_DIR = DIR  + identy +"/" # 临时目录
+if os.path.exists(TMP_DIR) is False:
+    os.makedirs(TMP_DIR)
+source_file = TMP_DIR + identy + "-tmp.txt"
+ts_file = TMP_DIR + identy + "-ts.txt"
+
 def handlerPrefix(url): # 解析前缀
     return "".join(url[:url.rfind("/")+1])
 
 def resolveM3u8(url,suffix=None,prefix=None): # 解析m3u8 获取ts列表
+    if os.path.exists(ts_file):
+        with open(ts_file,'r',encoding='utf-8') as f:
+            content = f.read()
+            return content.split("\n")
+    
     if prefix is None:
         prefix = handlerPrefix(url)
     
@@ -21,8 +44,7 @@ def resolveM3u8(url,suffix=None,prefix=None): # 解析m3u8 获取ts列表
     for line in content.split("\n"):
         if line.startswith("#EXT-X-KEY"):
             # 加密处理
-            isAes = True
-            print("视频已加密，暂时无法处理！")
+            print("视频已被加密，暂时无法处理！")
             return None
         if line.startswith("#") or line == '':
             continue
@@ -35,11 +57,13 @@ def resolveM3u8(url,suffix=None,prefix=None): # 解析m3u8 获取ts列表
     return list
 
 def generatorListFile(list,filename): # 格式化内容
+    if os.path.exists(file_name):
+        os.remove(file_name)
     with open(filename, "w") as f:
         for i in range(len(list)):
             file = "%d.ts" % i
             list[i]['filename'] = file
-            f.write("file " + DIR + file + "\n")
+            f.write("file " + TMP_DIR + file + "\n")
 
 def convertQueue(list): # 绑定queue
     qe = queue.Queue()
@@ -62,8 +86,10 @@ def downByThread(func,qe): # 多线程下载
     for t in threads:
         t.join()
 
-
 def downTs(url,filename): # 下载ts
+    if os.path.exists(filename):
+        print("file exists：%s" % filename)
+        return
     try:
         r = requests.get(url,stream=True)
         with open(filename, 'wb') as f:
@@ -81,7 +107,7 @@ def down(qe): # 执行下载逻辑
         url= ts['url']
         filename = ts['filename']
         try:
-            downTs(url,DIR + filename)
+            downTs(url,TMP_DIR + filename)
             print("Thread[%s] download %s is success[%s],use time[%s]!" % (thn,url,filename,datetime.datetime.now() - start))
         except Exception as e:
             print("[ERROR][DOWNLOAD]%s download is error!" % filename)
@@ -93,26 +119,28 @@ def merge(list_file,filename): # 合并ts为MP4
     command = "ffmpeg -y -f concat -safe 0 -i %s -vcodec copy %s" % (list_file,filename)
     print("exec video merge system command:",command)
     try:
-        os.system(command)
-        print("video merge is complate!")
+        # os.system(command)
+        import subprocess
+        p = subprocess.Popen(command, shell=True)
+        print("video merge is complate!",p.wait())
     except Exception as ex:
         print("video merge is error!")
         raise ex
 
 def delTsFile(): # 删除无用文件
-    for root, dirs, files in os.walk(DIR):
-        for file in files:
-            if file.endswith('.ts'):
-                os.remove(os.path.join(root, file))
+    # os.delete(TMP_DIR)
+    import shutil
+    shutil.rmtree(TMP_DIR)
+    # os.remove(TMP_DIR) # 无权删除，拒绝访问
 
 def repair(list): # 修复使用，下载缺失片段
     for item in list:
         url = item['url']
         filename = item['filename']
-        if os.path.exists(DIR + filename):
+        if os.path.exists(TMP_DIR + filename):
             continue
         try:
-            downTs(url,DIR + filename)
+            downTs(url,TMP_DIR + filename)
             print("download %s is success[%s]" % (url,filename))
         except Exception as e:
             print("[ERROR][DOWNLOAD]%s download is error!" % filename)
@@ -124,9 +152,9 @@ if __name__ == "__main__":
     if list:
         print("ready down ts count[%d]" % len(list))
         generatorListFile(list,source_file)
-        # repair(list)
         qe = convertQueue(list)
         downByThread(down,qe)
+        print("--- download complete,ready merge ts file ---")
         merge(source_file,target_file)
         delTsFile()
     print("use time[%s]!" % str(datetime.datetime.now() - start))
